backend/requirements.txt
# ── Core API + server ────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.27.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
loguru>=0.7.0
websockets>=12.0

# ── Data + indicators ────────────────────────────────────────────────────────
yfinance>=0.2.40
finnhub-python>=2.4.20
ta>=0.10.2
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=14.0.0

# ── Trading execution ────────────────────────────────────────────────────────
alpaca-py>=0.26.0

# ── ML / Deep learning ───────────────────────────────────────────────────────
# Install torch separately first: https://pytorch.org/get-started/locally/
torch>=2.2.0
transformers>=4.38.0
pytorch-forecasting>=1.1.0
lightning>=2.2.0
stable-baselines3>=2.3.0
gymnasium>=0.29.1

# ── Experiment tracking ──────────────────────────────────────────────────────
# greenlet is a transitive dep of mlflow/SQLAlchemy — install binary first:
#   pip install greenlet --only-binary :all:
greenlet>=3.0.0
mlflow>=2.10.0
optuna>=3.5.0
